Fix issue with returning a batch from Ahjo to proposal state
